feat(fs-utils): add utility functions for file existence checks and CSS entry candidates
